вынес методы в абстрактыней класс
